refactor(manager): 拆分 BinUpdater god class 为 BinSource 与英雄/地图处理器
Why:
- BinUpdater 单类 911 行，揉合 BIN 源抽象、英雄皮肤链路、地图公共去重链路与编排，职责过载，
  是核心数据层最大的结构集中点。

What:
- 新增 manager/bin_source.py（BinSource：WAD 优先/本地回退的 BIN 提取、地图 BIN 加载、base data）。
- 新增 manager/champion_bin_processor.py（ChampionBinProcessor：英雄皮肤链路）。
- 新增 manager/map_bin_processor.py（MapBinProcessor：地图链路）。
- BinUpdater 退化为编排器（read→filter→分派 processor），911 → 209 行；两个 processor 共享同一 BinSource。
- 顺带删除 _filter_data_by_ids 中恒为空的 languages 死字段（真实来源在 metadata.languages，下游无消费）。
- 改写 test_manager_bin_updater 白盒构造：原 BinUpdater.__new__ + 直调内部方法，改为针对
  BinSource/ChampionBinProcessor/MapBinProcessor 构造、patch 目标迁到方法所在模块；断言文案与期望全部不变。

Risk:
- 兼容性影响：无。BinUpdater 公共签名与 update() 行为、日志文案、产物结构、去重语义保持不变。
- 潜在风险：languages 由单一属性改为 update() 内向 bin_source/map_processor 同步；生产唯一入口
  update() 在分派前完成同步，无旁路。map 0 三次解析合并属行为敏感优化，保守跳过。

Test:
- uv run pytest -q：593 passed, 4 deselected（与重构前基准一致，主代理独立复跑确认）。
- ruff check 全绿；ruff format 仅作用于新文件，bin_updater.py 与测试无风格噪音。
