+++ tests/manager/test_manager_bin_updater.py
@@ -3,33 +3,37 @@ from types import SimpleNamespace
 
 import pytest
 
+from lol_audio_unpack.manager import bin_source as m_bin_source
 from lol_audio_unpack.manager import bin_updater as m_bin_updater
+from lol_audio_unpack.manager import champion_bin_processor as m_champion_processor
+from lol_audio_unpack.manager import map_bin_processor as m_map_processor
 from lol_audio_unpack.utils.run_summary import get_or_create_run_summary
 
 pytestmark = pytest.mark.unit
 
 
-def _build_updater(tmp_path: Path) -> m_bin_updater.BinUpdater:
-    """构造用于单测的最小 `BinUpdater` 实例。
+def _build_bin_source(tmp_path: Path) -> m_bin_source.BinSource:
+    """构造用于单测的最小 `BinSource` 实例。
 
     Args:
         tmp_path: pytest 提供的临时目录。
 
     Returns:
-        仅初始化当前测试所需字段的 `BinUpdater` 对象。
+        仅初始化当前测试所需字段的 `BinSource` 对象。
     """
-    updater = m_bin_updater.BinUpdater.__new__(m_bin_updater.BinUpdater)
-    updater.ctx = SimpleNamespace(config=SimpleNamespace(game_path=tmp_path, dev_mode=False), paths=SimpleNamespace())
-    updater.use_local_bin_flag_file = tmp_path / ".use_local_bin"
-    updater.local_bin_input_dir = tmp_path / "bin_input"
-    updater.local_bin_input_dir.mkdir(parents=True, exist_ok=True)
-    return updater
+    source = m_bin_source.BinSource.__new__(m_bin_source.BinSource)
+    source.ctx = SimpleNamespace(config=SimpleNamespace(game_path=tmp_path, dev_mode=False), paths=SimpleNamespace())
+    source.game_path = tmp_path
+    source.use_local_bin_flag_file = tmp_path / ".use_local_bin"
+    source.local_bin_input_dir = tmp_path / "bin_input"
+    source.local_bin_input_dir.mkdir(parents=True, exist_ok=True)
+    return source
 
 
 def test_extract_bin_raws_prefers_wad_when_wad_exists(tmp_path, monkeypatch):
     """验证存在 WAD 文件时优先从 WAD 提取 BIN 原始数据。"""
-    updater = _build_updater(tmp_path)
-    updater.use_local_bin_flag_file.write_text("", encoding="utf-8")
+    source = _build_bin_source(tmp_path)
+    source.use_local_bin_flag_file.write_text("", encoding="utf-8")
 
     calls = []
 
@@ -41,12 +45,12 @@ def test_extract_bin_raws_prefers_wad_when_wad_exists(tmp_path, monkeypatch):
             calls.append((self.path, list(bin_paths), raw))
             return [b"from-wad"]
 
-    monkeypatch.setattr(m_bin_updater, "WAD", FakeWAD)
+    monkeypatch.setattr(m_bin_source, "WAD", FakeWAD)
 
     wad_path = tmp_path / "Annie.wad.client"
     wad_path.write_bytes(b"")
 
-    result = updater._extract_bin_raws(
+    result = source._extract_bin_raws(
         wad_path=wad_path,
         bin_paths=["data/characters/Annie/skins/skin0001.bin"],
         entity_label="英雄 1 (annie)",
@@ -59,14 +63,14 @@ def test_extract_bin_raws_prefers_wad_when_wad_exists(tmp_path, monkeypatch):
 
 def test_extract_bin_raws_reads_local_files_when_flag_enabled(tmp_path):
     """验证启用本地 BIN 模式后会直接读取本地文件。"""
-    updater = _build_updater(tmp_path)
-    updater.use_local_bin_flag_file.write_text("", encoding="utf-8")
+    source = _build_bin_source(tmp_path)
+    source.use_local_bin_flag_file.write_text("", encoding="utf-8")
 
-    target_file = updater.local_bin_input_dir / "data/characters/Annie/skins/skin0001.bin"
+    target_file = source.local_bin_input_dir / "data/characters/Annie/skins/skin0001.bin"
     target_file.parent.mkdir(parents=True, exist_ok=True)
     target_file.write_bytes(b"from-local")
 
-    result = updater._extract_bin_raws(
+    result = source._extract_bin_raws(
         wad_path=tmp_path / "missing.wad.client",
         bin_paths=["data/characters/Annie/skins/skin0001.bin"],
         entity_label="英雄 1 (annie)",
@@ -78,17 +82,17 @@ def test_extract_bin_raws_reads_local_files_when_flag_enabled(tmp_path):
 
 def test_extract_bin_raws_logs_fallback_to_local_bin_when_wad_missing(tmp_path, monkeypatch):
     """验证 WAD 缺失但启用本地 BIN 模式时会输出回退诊断日志。"""
-    updater = _build_updater(tmp_path)
-    updater.use_local_bin_flag_file.write_text("", encoding="utf-8")
+    source = _build_bin_source(tmp_path)
+    source.use_local_bin_flag_file.write_text("", encoding="utf-8")
 
-    target_file = updater.local_bin_input_dir / "data/characters/Annie/skins/skin0001.bin"
+    target_file = source.local_bin_input_dir / "data/characters/Annie/skins/skin0001.bin"
     target_file.parent.mkdir(parents=True, exist_ok=True)
     target_file.write_bytes(b"from-local")
 
     debug_messages: list[str] = []
     trace_messages: list[str] = []
     monkeypatch.setattr(
-        m_bin_updater,
+        m_bin_source,
         "logger",
         SimpleNamespace(
             debug=lambda message: debug_messages.append(str(message)),
@@ -98,7 +102,7 @@ def test_extract_bin_raws_logs_fallback_to_local_bin_when_wad_missing(tmp_path, 
     )
 
     missing_wad = tmp_path / "missing.wad.client"
-    result = updater._extract_bin_raws(
+    result = source._extract_bin_raws(
         wad_path=missing_wad,
         bin_paths=["data/characters/Annie/skins/skin0001.bin"],
         entity_label="英雄 1 (annie)",
@@ -107,16 +111,16 @@ def test_extract_bin_raws_logs_fallback_to_local_bin_when_wad_missing(tmp_path, 
 
     assert result == [b"from-local"]
     assert debug_messages == [f"英雄 1 (annie) 的WAD文件不可用，回退到本地BIN模式: {missing_wad}"]
-    assert any(str(updater.local_bin_input_dir) in message for message in trace_messages)
+    assert any(str(source.local_bin_input_dir) in message for message in trace_messages)
 
 
 def test_extract_bin_raws_raises_when_flag_enabled_but_entity_dir_missing(tmp_path):
     """验证启用本地 BIN 模式但实体目录缺失时会抛出异常。"""
-    updater = _build_updater(tmp_path)
-    updater.use_local_bin_flag_file.write_text("", encoding="utf-8")
+    source = _build_bin_source(tmp_path)
+    source.use_local_bin_flag_file.write_text("", encoding="utf-8")
 
     with pytest.raises(FileNotFoundError, match="本地BIN实体目录不存在"):
-        updater._extract_bin_raws(
+        source._extract_bin_raws(
             wad_path=tmp_path / "missing.wad.client",
             bin_paths=["data/characters/Annie/skins/skin0001.bin"],
             entity_label="英雄 1 (annie)",
@@ -126,9 +130,9 @@ def test_extract_bin_raws_raises_when_flag_enabled_but_entity_dir_missing(tmp_pa
 
 def test_extract_bin_raws_returns_empty_when_local_mode_disabled(tmp_path):
     """验证未启用本地 BIN 模式时缺失 WAD 会返回空结果。"""
-    updater = _build_updater(tmp_path)
+    source = _build_bin_source(tmp_path)
 
-    result = updater._extract_bin_raws(
+    result = source._extract_bin_raws(
         wad_path=tmp_path / "missing.wad.client",
         bin_paths=["data/characters/Annie/skins/skin0001.bin"],
         entity_label="英雄 1 (annie)",
@@ -140,12 +144,12 @@ def test_extract_bin_raws_returns_empty_when_local_mode_disabled(tmp_path):
 
 def test_extract_bin_raws_raises_when_local_bin_root_missing(tmp_path):
     """验证本地 BIN 根目录缺失时会抛出异常。"""
-    updater = _build_updater(tmp_path)
-    updater.use_local_bin_flag_file.write_text("", encoding="utf-8")
-    updater.local_bin_input_dir.rmdir()
+    source = _build_bin_source(tmp_path)
+    source.use_local_bin_flag_file.write_text("", encoding="utf-8")
+    source.local_bin_input_dir.rmdir()
 
     with pytest.raises(FileNotFoundError, match="目录不存在"):
-        updater._extract_bin_raws(
+        source._extract_bin_raws(
             wad_path=tmp_path / "missing.wad.client",
             bin_paths=["data/characters/Annie/skins/skin0001.bin"],
             entity_label="英雄 1 (annie)",
@@ -155,11 +159,11 @@ def test_extract_bin_raws_raises_when_local_bin_root_missing(tmp_path):
 
 def test_extract_bin_raws_raises_for_path_traversal(tmp_path):
     """验证越界路径会被拒绝处理。"""
-    updater = _build_updater(tmp_path)
-    updater.use_local_bin_flag_file.write_text("", encoding="utf-8")
+    source = _build_bin_source(tmp_path)
+    source.use_local_bin_flag_file.write_text("", encoding="utf-8")
 
     with pytest.raises(ValueError, match="越界"):
-        updater._extract_bin_raws(
+        source._extract_bin_raws(
             wad_path=tmp_path / "missing.wad.client",
             bin_paths=["../../outside.bin"],
             entity_label="英雄 1 (annie)",
@@ -169,15 +173,15 @@ def test_extract_bin_raws_raises_for_path_traversal(tmp_path):
 
 def test_extract_bin_raws_allows_partial_missing_and_keeps_order(tmp_path):
     """验证部分 BIN 缺失时仍保持原始顺序返回结果。"""
-    updater = _build_updater(tmp_path)
-    updater.use_local_bin_flag_file.write_text("", encoding="utf-8")
+    source = _build_bin_source(tmp_path)
+    source.use_local_bin_flag_file.write_text("", encoding="utf-8")
 
-    base_dir = updater.local_bin_input_dir / "data/characters/Annie/skins"
+    base_dir = source.local_bin_input_dir / "data/characters/Annie/skins"
     base_dir.mkdir(parents=True, exist_ok=True)
     (base_dir / "skin0001.bin").write_bytes(b"first")
     (base_dir / "skin0003.bin").write_bytes(b"third")
 
-    result = updater._extract_bin_raws(
+    result = source._extract_bin_raws(
         wad_path=tmp_path / "missing.wad.client",
         bin_paths=[
             "data/characters/Annie/skins/skin0001.bin",
@@ -193,20 +197,20 @@ def test_extract_bin_raws_allows_partial_missing_and_keeps_order(tmp_path):
 
 def test_process_champion_skins_skips_when_first_bin_missing(tmp_path, monkeypatch):
     """验证首个皮肤 BIN 缺失时会跳过整组英雄皮肤处理。"""
-    updater = m_bin_updater.BinUpdater.__new__(m_bin_updater.BinUpdater)
-    updater.ctx = SimpleNamespace(config=SimpleNamespace(game_path=tmp_path, dev_mode=False), paths=SimpleNamespace())
-    updater.force_update = False
-    updater.process_events = False
-    updater.version = "16.3"
-    updater.game_path = tmp_path
-    updater.champion_banks_dir = tmp_path / "banks" / "champions"
-    updater.champion_events_dir = tmp_path / "events" / "champions"
+    processor = m_champion_processor.ChampionBinProcessor.__new__(m_champion_processor.ChampionBinProcessor)
+    processor.ctx = SimpleNamespace(config=SimpleNamespace(game_path=tmp_path, dev_mode=False), paths=SimpleNamespace())
+    processor.force_update = False
+    processor.process_events = False
+    processor.version = "16.3"
+    processor.game_path = tmp_path
+    processor.champion_banks_dir = tmp_path / "banks" / "champions"
+    processor.champion_events_dir = tmp_path / "events" / "champions"
+    processor.bin_source = SimpleNamespace(_extract_bin_raws=lambda *args, **kwargs: [None, b"fallback"])
 
-    monkeypatch.setattr(m_bin_updater, "needs_update", lambda *args, **kwargs: True)
-    monkeypatch.setattr(updater, "_extract_bin_raws", lambda *args, **kwargs: [None, b"fallback"])
+    monkeypatch.setattr(m_champion_processor, "needs_update", lambda *args, **kwargs: True)
 
     write_calls = []
-    monkeypatch.setattr(m_bin_updater, "write_data", lambda *args, **kwargs: write_calls.append(True))
+    monkeypatch.setattr(m_champion_processor, "write_data", lambda *args, **kwargs: write_calls.append(True))
 
     champion_data = {
         "alias": "Annie",
@@ -217,18 +221,18 @@ def test_process_champion_skins_skips_when_first_bin_missing(tmp_path, monkeypat
         "wad": {"root": "Game/DATA/FINAL/Champions/Annie.wad.client"},
     }
 
-    updater._process_champion_skins(champion_data, "1")
+    processor._process_champion_skins(champion_data, "1")
 
     assert write_calls == []
 
 
 def test_load_map_bin_file_reads_local_bin_when_available(tmp_path, monkeypatch):
     """验证地图 BIN 可用时优先读取本地文件内容。"""
-    updater = _build_updater(tmp_path)
-    updater.game_path = tmp_path
-    updater.use_local_bin_flag_file.write_text("", encoding="utf-8")
+    source = _build_bin_source(tmp_path)
+    source.game_path = tmp_path
+    source.use_local_bin_flag_file.write_text("", encoding="utf-8")
 
-    local_file = updater.local_bin_input_dir / "data/maps/shipping/map11/map11.bin"
+    local_file = source.local_bin_input_dir / "data/maps/shipping/map11/map11.bin"
     local_file.parent.mkdir(parents=True, exist_ok=True)
     local_file.write_bytes(b"map-bin")
 
@@ -236,9 +240,9 @@ def test_load_map_bin_file_reads_local_bin_when_available(tmp_path, monkeypatch)
         def __init__(self, raw):
             self.raw = raw
 
-    monkeypatch.setattr(m_bin_updater, "BIN", FakeBIN)
+    monkeypatch.setattr(m_bin_source, "BIN", FakeBIN)
 
-    result = updater._load_map_bin_file(
+    result = source._load_map_bin_file(
         "11",
         {
             "binPath": "data/maps/shipping/map11/map11.bin",
@@ -258,14 +262,15 @@ def test_update_records_note_when_targeted_maps_exclude_common_map(tmp_path, mon
     updater.process_events = True
     updater.version = "16.3"
     updater.data_file_base = tmp_path / "data"
-    updater.use_local_bin_flag_file = tmp_path / ".use_local_bin"
+    updater.languages = []
+    updater.bin_source = SimpleNamespace(languages=[], _is_local_bin_mode_enabled=lambda: False)
+    updater._map_processor = SimpleNamespace(languages=[], _update_maps=lambda data: data)
 
     monkeypatch.setattr(
         m_bin_updater,
         "read_data",
         lambda *args, **kwargs: {"metadata": {"languages": ["zh_CN"]}, "maps": {"33": {"id": 33}}},
     )
-    monkeypatch.setattr(updater, "_update_maps", lambda data: data)
 
     updater.update(target="map", map_ids=["33"])
 
@@ -282,6 +287,9 @@ def test_update_logs_stage_start_and_summary_for_targeted_mode(tmp_path, monkeyp
     updater.version = "16.3"
     updater.data_file_base = tmp_path / "data"
     updater.languages = []
+    updater.bin_source = SimpleNamespace(languages=[], _is_local_bin_mode_enabled=lambda: True)
+    updater._champion_processor = SimpleNamespace(_update_champions=lambda _data: None)
+    updater._map_processor = SimpleNamespace(languages=[], _update_maps=lambda _data: None)
 
     info_messages: list[str] = []
     success_messages: list[str] = []
@@ -303,10 +311,7 @@ def test_update_logs_stage_start_and_summary_for_targeted_mode(tmp_path, monkeyp
             success=lambda message: success_messages.append(str(message)),
         ),
     )
-    monkeypatch.setattr(updater, "_is_local_bin_mode_enabled", lambda: True)
-    monkeypatch.setattr(updater, "_update_champions", lambda _data: None)
     monkeypatch.setattr(updater, "_record_map_event_scope_note", lambda _map_ids: None)
-    monkeypatch.setattr(updater, "_update_maps", lambda _data: None)
 
     updater.update(target="all", champion_ids=["1"], map_ids=["11"])
 
@@ -316,19 +321,18 @@ def test_update_logs_stage_start_and_summary_for_targeted_mode(tmp_path, monkeyp
 
 def test_process_single_map_records_note_when_common_dedup_removes_all_events(tmp_path, monkeypatch):
     """验证公共事件去重清空结果时会记录可解释差异。"""
-    updater = m_bin_updater.BinUpdater.__new__(m_bin_updater.BinUpdater)
-    updater.ctx = SimpleNamespace(
+    processor = m_map_processor.MapBinProcessor.__new__(m_map_processor.MapBinProcessor)
+    processor.ctx = SimpleNamespace(
         config=SimpleNamespace(game_path=tmp_path, dev_mode=False),
         runtime_cache={},
         paths=SimpleNamespace(),
     )
-    updater.force_update = False
-    updater.process_events = True
-    updater.version = "16.3"
-    updater.game_path = tmp_path
-    updater.languages = []
-    updater.map_banks_dir = tmp_path / "banks" / "maps"
-    updater.map_events_dir = tmp_path / "events" / "maps"
+    processor.force_update = False
+    processor.process_events = True
+    processor.version = "16.3"
+    processor.languages = []
+    processor.map_banks_dir = tmp_path / "banks" / "maps"
+    processor.map_events_dir = tmp_path / "events" / "maps"
 
     fake_bin = SimpleNamespace(
         theme_music=None,
@@ -346,11 +350,12 @@ def test_process_single_map_records_note_when_common_dedup_removes_all_events(tm
         ],
     )
 
-    monkeypatch.setattr(m_bin_updater, "needs_update", lambda *args, **kwargs: True)
-    monkeypatch.setattr(m_bin_updater, "write_data", lambda *args, **kwargs: None)
-    monkeypatch.setattr(updater, "_load_map_bin_file", lambda *_args, **_kwargs: fake_bin)
+    processor.bin_source = SimpleNamespace(_load_map_bin_file=lambda *_args, **_kwargs: fake_bin)
 
-    updater._process_single_map(
+    monkeypatch.setattr(m_map_processor, "needs_update", lambda *args, **kwargs: True)
+    monkeypatch.setattr(m_map_processor, "write_data", lambda *args, **kwargs: None)
+
+    processor._process_single_map(
         "33",
         {
             "binPath": "data/maps/shipping/map33/map33.bin",
@@ -360,7 +365,7 @@ def test_process_single_map_records_note_when_common_dedup_removes_all_events(tm
         set(),
     )
 
-    summary = get_or_create_run_summary(updater.ctx.runtime_cache)
+    summary = get_or_create_run_summary(processor.ctx.runtime_cache)
     assert any(
         "地图 33 (Map33) 的事件在与 地图 0 的公共事件去重后为空" in note for note in summary.stages["update"].notes
     )
@@ -369,20 +374,20 @@ def test_process_single_map_records_note_when_common_dedup_removes_all_events(tm
 
 def test_update_champions_logs_simple_progress_messages(tmp_path, monkeypatch):
     """验证英雄批量更新会输出简单的日志进度提示。"""
-    updater = m_bin_updater.BinUpdater.__new__(m_bin_updater.BinUpdater)
-    updater.ctx = SimpleNamespace(
+    processor = m_champion_processor.ChampionBinProcessor.__new__(m_champion_processor.ChampionBinProcessor)
+    processor.ctx = SimpleNamespace(
         config=SimpleNamespace(game_path=tmp_path, dev_mode=False),
         runtime_cache={},
         paths=SimpleNamespace(),
     )
-    updater.champion_banks_dir = tmp_path / "banks" / "champions"
-    updater.champion_events_dir = tmp_path / "events" / "champions"
+    processor.champion_banks_dir = tmp_path / "banks" / "champions"
+    processor.champion_events_dir = tmp_path / "events" / "champions"
     processed_ids: list[str] = []
     info_messages: list[str] = []
     success_messages: list[str] = []
 
     monkeypatch.setattr(
-        m_bin_updater,
+        m_champion_processor,
         "logger",
         SimpleNamespace(
             info=lambda message: info_messages.append(str(message)),
@@ -390,12 +395,12 @@ def test_update_champions_logs_simple_progress_messages(tmp_path, monkeypatch):
         ),
     )
     monkeypatch.setattr(
-        updater,
+        processor,
         "_process_champion_skins",
         lambda _champion_data, champion_id: processed_ids.append(champion_id),
     )
 
-    updater._update_champions({"champions": {"2": {}, "1": {}}})
+    processor._update_champions({"champions": {"2": {}, "1": {}}})
 
     assert processed_ids == ["1", "2"]
     assert "处理英雄进度 1/2: 1" in info_messages
@@ -405,20 +410,20 @@ def test_update_champions_logs_simple_progress_messages(tmp_path, monkeypatch):
 
 def test_update_maps_logs_simple_progress_messages(tmp_path, monkeypatch):
     """验证地图批量更新会输出简单的日志进度提示。"""
-    updater = m_bin_updater.BinUpdater.__new__(m_bin_updater.BinUpdater)
-    updater.ctx = SimpleNamespace(
+    processor = m_map_processor.MapBinProcessor.__new__(m_map_processor.MapBinProcessor)
+    processor.ctx = SimpleNamespace(
         config=SimpleNamespace(game_path=tmp_path, dev_mode=False),
         runtime_cache={},
         paths=SimpleNamespace(),
     )
-    updater.map_banks_dir = tmp_path / "banks" / "maps"
-    updater.map_events_dir = tmp_path / "events" / "maps"
+    processor.map_banks_dir = tmp_path / "banks" / "maps"
+    processor.map_events_dir = tmp_path / "events" / "maps"
     processed_ids: list[str] = []
     info_messages: list[str] = []
     success_messages: list[str] = []
 
     monkeypatch.setattr(
-        m_bin_updater,
+        m_map_processor,
         "logger",
         SimpleNamespace(
             info=lambda message: info_messages.append(str(message)),
@@ -428,12 +433,12 @@ def test_update_maps_logs_simple_progress_messages(tmp_path, monkeypatch):
         ),
     )
     monkeypatch.setattr(
-        updater,
+        processor,
         "_process_single_map",
         lambda map_id, *_args: processed_ids.append(map_id),
     )
 
-    updater._update_maps({"maps": {"11": {"name": "Map11"}, "12": {"name": "Map12"}}})
+    processor._update_maps({"maps": {"11": {"name": "Map11"}, "12": {"name": "Map12"}}})
 
     assert processed_ids == ["11", "12"]
     assert "处理地图进度 1/2: 11" in info_messages
